Fix QAssert for non-marmalade builds.
CCAssert is not variadic so assert message need to be foramtted
before being passwd to QAssert.  Making this work transparently
in a cross platform way is likely to be hard since varadic macros
are have issues on some platforms.

Also CCAssert must be a macro so the underlying assert call can
get the correct line/function/file information, so the marmalade
one will need to be changed to a macro too.