Quick app wasn't paused after suspend
On suspend/resume sequence, despite using the documented system:pauseTimers()/
resumeTimers(), the time skipped on the app clock as if suspend/resume had not occurred.
Turned out these calls did very little, the suspend seemed to work at least partly because
cocos2d was itself suspended anyway. However, the first "update" time was for the real-time
delta since the last time. Changed behaviour so that on resumeTimers(), the first update
is ignored - as just in case are any when paused. Added internal property _stalledTimer
to achieve this.

Claim Fixed: MSDK-2169
Reviewed by: abaude.