MSDK-2702 Remove requirement to copy quicklua directory into each Quick application
(Private) config setting [quick]QuickLuaDir is used to pass where the quicklua directory is. Only
expected to be used in simulator. Default is quicklua in the assets area, as before.

Fix LuaLoadFile. i) return -1 if a bad filename (status==0 to show valid response, it is not boolean).
ii) Check the file exists directly rather than relying on MainLuaLoadData not working - for some reason
the latter appears to work when it should not.

Claim Fixed: MSDK-2702