MSDK-2702 Remove requirement to copy quicklua directory into each Quick application
Add feature so we can supply the original filename when precompiling, so that
when we expand to a fullpath we still have the original to create the proper destination.

Only support QUICKLUA environment variable on simulator - elsewhere we expect all
the files to be copied to the target anyway and don't need the feature.